pyproject.toml
[tool.poetry]
name = "crypto-trading-system"
version = "0.1.0"
description = "Cryptocurrency trading monitoring and prediction system"
authors = ["Devin"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
python-binance = "^1.0.25"
pandas = "^2.2.3"
numpy = "^2.2.0"
scikit-learn = "^1.6.0"
nltk = "^3.9.1"
snownlp = "^0.12.3"
easyocr = "^1.7.2"
pillow = "^11.0.0"
requests = "^2.32.3"
python-multipart = "^0.0.20"
aiosmtplib = "^3.0.2"
jinja2 = "^3.1.4"
python-dotenv = "^1.0.1"
xgboost = "^2.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
